🎨 style(index.html): update body class to use flexbox and min-height The body class has been updated to use flexbox and min-height to ensure that the content is always visible on the screen. This improves the user experience by preventing content from being cut off or hidden.
🐛 fix(Todo.jsx): add start and end date to new todo object
✨ feat(Todo.jsx): add start and end date input fields to form
The new todo object was missing the start and end date properties. This has been fixed by adding the start and end date properties to the new todo object. Two new input fields have been added to the form to allow users to input the start and end dates for the new todo.

🎨 style(Todo.jsx): improve the UI of the To Do list
The To Do list UI has been improved by adding a table to display the To Do items. The table has columns for Done?, Description, Start Date, End Date, and Delete. The Description column now has a strikethrough for completed items.