🎨 style(index.html): remove unused script tag 💄 style(Calendar.jsx): improve styling and readability The script tag in index.html was removed as it was not being used. In Calendar.jsx, the styling was improved by removing unnecessary classes and adding z-index to the container. The function `capitalizeFirstLetter` was imported from a separate file to improve readability and maintainability. The function name was also updated to better reflect its purpose.
🎨 style(Header.jsx, Footer.jsx): adjust styling of header and footer components
The changes made to the Header component include reducing the margin-bottom from 14 to 10, and removing the unnecessary text-align property. The changes made to the Footer component include increasing the margin-top from 20 to 60, which improves the spacing between the content and the footer.

🎨 style(Table.jsx): simplify the title of the table
🔄 refactor(functions.js): rename functions.ts to functions.js
The title of the table has been simplified to either "Create a To Do" or "List of To Do's" depending on whether there are any ToDos. The file functions.ts has been renamed to functions.js to match the file extension.

🔥 refactor(Todo.jsx): remove unused code
The console.log statements and unused functions have been removed from the code. This improves the code's readability and maintainability.