🚀 chore(package.json): add react-big-calendar and sass dependencies This commit adds the react-big-calendar package as a dependency to the project. It also adds the node-sass and sass-loader packages as dev dependencies to enable the use of Sass in the project.
🔨 refactor(Calendar.jsx): replace custom calendar component with react-big-calendar
The custom calendar component has been replaced with the react-big-calendar library. This library provides a more robust and customizable calendar component with support for events and date localization. The new component is named MyCalendar and takes in an array of events as props.

🐛 fix(Table.jsx): change property names from startDate and endDate to start and end
✨ feat(Table.jsx): change property name from text to title
The property names for start and end dates were changed from startDate and endDate to start and end, respectively, to improve consistency with the naming conventions. The property name for the todo text was changed to title to better reflect the content of the property.

🎨 style(Todo.jsx): format code and add whitespace
✨ feat(Todo.jsx): add react-big-calendar component
The code has been formatted to improve readability. The react-big-calendar component has been added to the TodoList component. The component displays the ToDos in a calendar view, which improves the user experience by providing an alternative view of the ToDos.

🔨 refactor(test.jsx): remove TODO comments and unused code
✨ feat(test.jsx): add CalendarPage component
The TODO comments and unused code have been removed to improve code readability. The CalendarPage component has been added, which uses the react-big-calendar library to display events stored in local storage. The component uses the useState and useEffect hooks to manage the state of the calendarItems array, which is stored in local storage. The useEffect hook is used to load the calendarItems array from local storage when the component mounts, and to save the array to local storage when it changes. The component also uses the moment and momentLocalizer libraries to format dates for display in the calendar.